refactor: Waiting 도메인 분리
