+++ src/main/java/nextstep/reservation/ReservationService.java
@@ -9,14 +9,12 @@ import nextstep.exception.dataaccess.DataAccessErrorCode;
 import nextstep.exception.dataaccess.DataAccessException;
 import nextstep.member.Member;
 import nextstep.member.MemberDao;
-import nextstep.reservation.dto.response.ReservationWaitingResponseDto;
 import nextstep.schedule.Schedule;
 import nextstep.schedule.ScheduleDao;
 import nextstep.theme.Theme;
 import nextstep.theme.ThemeDao;
 import org.springframework.stereotype.Service;
 
-import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
 import java.util.stream.Collectors;
@@ -70,25 +68,6 @@ public class ReservationService {
         reservationDao.deleteById(id);
     }
 
-    public Long createWaiting(Member member, Long scheduleId) {
-        if (member == null) {
-            throw new AuthException(AuthErrorCode.INVALID_USER);
-        }
-        Schedule schedule = scheduleDao.findById(scheduleId)
-                .orElseThrow(() -> new DataAccessException(DataAccessErrorCode.SCHEDULE_NOT_FOUND));
-
-        Reservation newReservation = new Reservation(
-                schedule,
-                member
-        );
-
-        return reservationDao.save(newReservation);
-    }
-
-    public void deleteWaitingById(Member member, Long id) {
-        deleteById(member, id);
-    }
-
     public List<Reservation> getReservationsByMember(Member member) {
         if (member == null) {
             throw new AuthException(AuthErrorCode.INVALID_USER);
@@ -101,21 +80,6 @@ public class ReservationService {
                     return !isWaitingReservation(reservationDao.getPriority(scheduleId, waitTicketNumber));
                 })
                 .collect(Collectors.toList());
-    }
-
-    public List<ReservationWaitingResponseDto> getReservationWaitingsByMember(Member member) {
-        if (member == null) {
-            throw new AuthException(AuthErrorCode.INVALID_USER);
-        }
-        List<Reservation> reservationsByMemberId = reservationDao.findByMemberId(member.getId());
-        List<ReservationWaitingResponseDto> result = new ArrayList<>();
-        for (Reservation reservation : reservationsByMemberId) {
-            Long waitNum = reservationDao.getPriority(reservation.getSchedule().getId(), reservation.getWaitTicketNumber());
-            if (isWaitingReservation(waitNum)) {
-                result.add(ReservationWaitingResponseDto.of(reservation, waitNum));
-            }
-        }
-        return result;
     }
 
     private static boolean isWaitingReservation(Long waitNum) {
